moved clinical data filtering methods into separate classes
