+++ web/src/main/java/org/cbioportal/web/util/ClinicalDataBinner.java
@@ -4,14 +4,22 @@ import org.cbioportal.model.ClinicalData;
 import org.cbioportal.model.ClinicalDataBinCount;
 import org.springframework.stereotype.Component;
 
+import java.util.Collections;
 import java.util.List;
 
 @Component
 public class ClinicalDataBinner
 {
-    public List<ClinicalDataBinCount> calculateDataBins(List<ClinicalData> clinicalData)
+    public List<ClinicalDataBinCount> calculateDataBins(String attributeId, List<ClinicalData> clinicalData)
     {
         // TODO properly calculate data bins wrt clinical attribute
-        return null;
+        
+        ClinicalDataBinCount count = new ClinicalDataBinCount();
+        
+        count.setAttributeId(attributeId);
+        count.setCount(clinicalData.size());
+        count.setValue("ALL");
+        
+        return Collections.singletonList(count);
     }
 }
